Relayer PositronConsoleService from common to browser
