+++ src/vs/workbench/contrib/languageRuntime/browser/languageRuntime.contribution.ts
@@ -2,7 +2,7 @@
  *  Copyright (C) 2022 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
 
-import { registerLanguageRuntimeActions } from 'vs/workbench/contrib/languageRuntime/common/languageRuntimeActions';
+import { registerLanguageRuntimeActions } from 'vs/workbench/contrib/languageRuntime/browser/languageRuntimeActions';
 
 // Register commands
 registerLanguageRuntimeActions();
