+++ src/vs/workbench/contrib/positronConsole/browser/components/activityErrorStream.tsx
@@ -5,7 +5,7 @@
 import 'vs/css!./activityErrorStream';
 import * as React from 'react';
 import { OutputLines } from 'vs/workbench/contrib/positronConsole/browser/components/outputLines';
-import { ActivityItemErrorStream } from 'vs/workbench/services/positronConsole/common/classes/activityItemStream';
+import { ActivityItemErrorStream } from 'vs/workbench/services/positronConsole/browser/classes/activityItemStream';
 
 // ActivityErrorStreamProps interface.
 export interface ActivityErrorStreamProps {
