+++ src/vs/workbench/contrib/positronConsole/browser/components/runtimeStartup.tsx
@@ -5,7 +5,7 @@
 import 'vs/css!./runtimeStartup';
 import * as React from 'react';
 import { OutputLines } from 'vs/workbench/contrib/positronConsole/browser/components/outputLines';
-import { RuntimeItemStartup } from 'vs/workbench/services/positronConsole/common/classes/runtimeItemStartup';
+import { RuntimeItemStartup } from 'vs/workbench/services/positronConsole/browser/classes/runtimeItemStartup';
 
 // RuntimeStartupProps interface.
 export interface RuntimeStartupProps {
