+++ src/vs/workbench/contrib/positronConsole/browser/positronConsoleActions.ts
@@ -28,7 +28,7 @@ import { NOTEBOOK_EDITOR_FOCUSED } from 'vs/workbench/contrib/notebook/common/no
 import { PositronConsoleViewPane } from 'vs/workbench/contrib/positronConsole/browser/positronConsoleView';
 import { confirmationModalDialog } from 'vs/workbench/browser/positronModalDialogs/confirmationModalDialog';
 import { IExecutionHistoryService } from 'vs/workbench/contrib/executionHistory/common/executionHistoryService';
-import { IPositronConsoleService, POSITRON_CONSOLE_VIEW_ID } from 'vs/workbench/services/positronConsole/common/interfaces/positronConsoleService';
+import { IPositronConsoleService, POSITRON_CONSOLE_VIEW_ID } from 'vs/workbench/services/positronConsole/browser/interfaces/positronConsoleService';
 import { isString } from 'vs/base/common/types';
 
 /**
