+++ src/vs/workbench/services/positronConsole/browser/classes/activityItemOutputHtml.ts
