+++ src/vs/workbench/services/positronConsole/browser/classes/activityItemOutputPlot.ts
