+++ src/vs/workbench/services/positronConsole/browser/classes/activityItemStream.ts
