+++ src/vs/workbench/workbench.common.main.ts
@@ -420,7 +420,7 @@ import 'vs/workbench/contrib/executionHistory/common/executionHistory';
 // Workbench services
 import 'vs/workbench/services/languageRuntime/common/languageRuntime';
 import 'vs/workbench/contrib/positronHelp/browser/positronHelpService';
-import 'vs/workbench/services/positronConsole/common/positronConsoleService';
+import 'vs/workbench/services/positronConsole/browser/positronConsoleService';
 import 'vs/workbench/services/positronEnvironment/common/positronEnvironmentService';
 
 // --- End Positron ---
